Fix date + Int fields to floats

+++ api_tracability/models.py
@@ -16,24 +16,24 @@ class MiseEnCulture(Action):
 
 
 class NourrissageHumide(Action):
-    given_quantity = models.IntegerField()
-    given_quantity_bac = models.IntegerField()
+    given_quantity = models.FloatField()
+    given_quantity_bac = models.FloatField()
     marc_arrival_date = models.DateTimeField()
     anomaly = models.BooleanField()
     anomaly_comment = models.TextField(blank=True, null=True)
     is_imw100_weighted = models.BooleanField()
-    imw100_weight = models.IntegerField(blank=True, null=True)
+    imw100_weight = models.FloatField(blank=True, null=True)
 
 
 class NourrissageSon(Action):
-    given_quantity = models.IntegerField()
-    given_quantity_bac = models.IntegerField()
+    given_quantity = models.FloatField()
+    given_quantity_bac = models.FloatField()
     son_arrival_date = models.DateTimeField()
 
 
 class Tamisage(Action):
-    sieved_quantity = models.IntegerField()
+    sieved_quantity = models.FloatField()
 
 
 class Recolte(Action):
-    harvested_quantity = models.IntegerField()
+    harvested_quantity = models.FloatField()
